requirements.txt
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Werkzeug==1.0.1
Pandas==2.1.0
sqlalchemy==2.0.20
psycopg2==2.8.6
python-dotenv==1.0.0